fix(update): isolate finalization types from shared primitives
